requirements.txt
Flask>=2.0
Flask-WTF>=1.0
SQLAlchemy>=1.4
Flask-SQLAlchemy>=3.0
WTForms>=3.0
python-dotenv>=1.0